requirements.txt
Django
djangorestframework
pytest
pytest-django
pytest-pythonpath
flake8
coverage
